FreeWiFi/miniaqua/requirements.txt
aiohttp==3.8.5
psycopg2==2.9.9
python-dotenv==1.0.1
wheel==0.37.0
httpx==0.25.0
pykakasi==2.2.1
colorlog==6.8.2
pypresence==4.3.0
tweepy==4.14.0
requests-oauthlib==1.3.1